+++ src/ResourceManager/RecoveryServices.SiteRecovery/Commands.RecoveryServices.SiteRecovery/Vault/ImportAzureRmRecoveryServicesAsrVaultSettingsFile.cs
@@ -1,4 +1,4 @@
-// ----------------------------------------------------------------------------------
+﻿// ----------------------------------------------------------------------------------
 // 
 // Copyright Microsoft Corporation
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -27,10 +27,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
     ///     Imports the specified ASR vault settings file to set the vault context(PowerShell session context)
     ///     for subsequent ASR operations in the PowerShell session. 
     /// </summary>
-    [Cmdlet(
-        VerbsData.Import,
-        "AzureRmRecoveryServicesAsrVaultSettingsFile",
-        SupportsShouldProcess = true)]
+    [Cmdlet("Import", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "RecoveryServicesAsrVaultSettingsFile",SupportsShouldProcess = true)]
     [OutputType(typeof(ASRVaultSettings))]
     [Alias(
         "Import-ASRVaultSettingsFile")]