+++ src/ResourceManager/Relay/Commands.Relay/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fixed issue with default resource groups not being set.
+* Updated to the latest version of the Azure ClientRuntime.
 
 ## Version 0.3.7
 * Updated help files to include full parameter types and correct input/output types.
@@ -44,7 +46,7 @@
 ## Version 0.3.0
 * Add support for online help
     - Run Get-Help with the -Online parameter to open the online help in your default Internet browser
-    
+
 ## Version 0.2.7
 
 ## Version 0.2.6